requirements.txt
torch
torchvision
scikit-learn
numpy
pandas
matplotlib
scipy
jupyter
tqdm
umap-learn
seaborn